+++ GRDB/Utils/Pool.swift
@@ -47,7 +47,7 @@ final class Pool<T> {
     }
     
     private let makeElement: () throws -> T
-    private var items: ReadWriteBox<[Item]> = ReadWriteBox(value: [])
+    @ReadWriteBox private var items: [Item] = []
     private let itemsSemaphore: DispatchSemaphore // limits the number of elements
     private let itemsGroup: DispatchGroup         // knows when no element is used
     private let barrierQueue: DispatchQueue
@@ -67,7 +67,7 @@ final class Pool<T> {
             itemsSemaphore.wait()
             itemsGroup.enter()
             do {
-                let item = try items.write { items -> Item in
+                let item = try $items.update { items -> Item in
                     if let item = items.first(where: \.isAvailable) {
                         item.isAvailable = false
                         return item
@@ -96,7 +96,7 @@ final class Pool<T> {
     }
     
     private func release(_ item: Item) {
-        items.write { _ in
+        $items.update { _ in
             // This is why Item is a class, not a struct: so that we can
             // release it without having to find in it the items array.
             item.isAvailable = true
@@ -108,7 +108,7 @@ final class Pool<T> {
     /// Performs a block on each pool element, available or not.
     /// The block is run is some arbitrary dispatch queue.
     func forEach(_ body: (T) throws -> Void) rethrows {
-        try items.read { items in
+        try $items.read { items in
             for item in items {
                 try body(item.element)
             }
@@ -118,9 +118,7 @@ final class Pool<T> {
     /// Removes all elements from the pool.
     /// Currently used elements won't be reused.
     func removeAll() {
-        items.write {
-            $0.removeAll()
-        }
+        items = []
     }
     
     /// Blocks until no element is used, and runs the `barrier` function before