+++ Tests/GRDBTests/FTS5WrapperTokenizerTests.swift
@@ -37,7 +37,7 @@ private final class LatinAsciiTokenizer : FTS5WrapperTokenizer {
     
     func accept(token: String, flags: FTS5TokenFlags, for tokenization: FTS5Tokenization, tokenCallback: FTS5WrapperTokenCallback) throws {
         // Convert token to Latin-ASCII and lowercase
-        if #available(iOS 9.0, OSX 10.11, *) {
+        if #available(OSX 10.11, *) {
             if let token = token.applyingTransform(StringTransform("Latin-ASCII; Lower"), reverse: false) {
                 try tokenCallback(token, flags)
             }